# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: OpenLoadBalancer Admin API
  description: |
    REST API for managing OpenLoadBalancer instances.

    ## Authentication
    All endpoints require Bearer token authentication via the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    API requests are rate-limited to 1000 requests per minute per IP.
  version: 0.1.0
  contact:
    name: OpenLoadBalancer Team
    url: https://openloadbalancer.dev
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8081
    description: Local development
  - url: https://api.openloadbalancer.dev
    description: Production
...
paths:
...
  /api/v1/pools:
    get:
      tags: [Pools]
      summary: List pools
      description: Returns all backend pools with their backends and health info
      operationId: listPools
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: List of pools
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/pools/{name}:
    get:
      tags: [Pools]
      summary: Get pool details
      description: Returns detailed information for a specific pool
      operationId: getPool
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/PoolName'
      responses:
        '200':
          description: Pool details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolInfoResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ---------------------------------------------------------------------------
  # Backends
  # ---------------------------------------------------------------------------
  /api/v1/backends:
    get:
      tags: [Backends]
      summary: List pool names
      description: Returns the names of all backend pools
      operationId: listBackendPools
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: List of pool names
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackendPoolListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/backends/{pool}:
    get:
      tags: [Backends]
      summary: List backends in pool
      description: Returns all backends in a specific pool
      operationId: listBackendsInPool
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/PoolName'
      responses:
        '200':
          description: List of backends
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackendListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Backends]
      summary: Add backend to pool
      description: Adds a new backend to the specified pool (Raft-aware in clustered mode)
      operationId: addBackend
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/PoolName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddBackendRequest'
      responses:
        '201':
          description: Backend added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/backends/{pool}/{backend}:
    get:
      tags: [Backends]
      summary: Get backend details
      description: Returns detailed information for a specific backend
      operationId: getBackend
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/PoolName'
        - $ref: '#/components/parameters/BackendID'
      responses:
        '200':
          description: Backend details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackendInfoResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [Backends]
      summary: Update backend
      description: Updates backend weight and/or max connections (Raft-aware in clustered mode)
      operationId: updateBackend
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/PoolName'
        - $ref: '#/components/parameters/BackendID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBackendRequest'
      responses:
        '200':
          description: Backend updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Backends]
      summary: Remove backend
      description: Removes a backend from the pool (Raft-aware in clustered mode)
      operationId: removeBackend
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/PoolName'
        - $ref: '#/components/parameters/BackendID'
      responses:
        '200':
          description: Backend removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/backends/{pool}/{backend}/drain:
    post:
      tags: [Backends]
      summary: Drain backend
      description: Sets the backend to draining state — stops sending new connections while in-flight requests complete
      operationId: drainBackend
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/PoolName'
        - $ref: '#/components/parameters/BackendID'
      responses:
        '200':
          description: Backend draining
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ---------------------------------------------------------------------------
  # Routes
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PoolName:
      name: pool
      in: path
      required: true
      description: Backend pool name
      schema:
        type: string
    BackendID:
      name: backend
      in: path
      required: true
      description: Backend identifier
      schema:
        type: string
  responses:
    Unauthorized:
      description: Unauthorized — missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string

    # -------------------------------------------------------------------------
    # System
    # -------------------------------------------------------------------------
...
    PoolInfo:
      type: object
      properties:
        name:
          type: string
        algorithm:
          type: string
        backends:
          type: array
          items:
            $ref: '#/components/schemas/BackendInfo'
        total:
          type: integer
        healthy:
          type: integer
        health_check:
          $ref: '#/components/schemas/HealthCheckInfo'
    HealthCheckInfo:
      type: object
      properties:
        enabled:
          type: boolean
        interval:
          type: integer
          description: Interval in nanoseconds
        timeout:
          type: integer
          description: Timeout in nanoseconds
        path:
          type: string
        port:
          type: integer
    PoolListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/PoolInfo'
    PoolInfoResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PoolInfo'

    # -------------------------------------------------------------------------
    # Backends
    # -------------------------------------------------------------------------
    BackendInfo:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
        weight:
          type: integer
          format: int32
        max_conns:
          type: integer
          format: int32
        state:
          type: string
          enum: [up, down, draining, starting]
        healthy:
          type: boolean
        active_conns:
          type: integer
          format: int64
        total_requests:
          type: integer
          format: int64
        total_errors:
          type: integer
          format: int64
        total_bytes:
          type: integer
          format: int64
        avg_latency:
          type: string
        last_latency:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
    AddBackendRequest:
      type: object
      required: [id, address]
      properties:
        id:
          type: string
        address:
          type: string
        weight:
          type: integer
          minimum: 0
    UpdateBackendRequest:
      type: object
      properties:
        weight:
          type: integer
          format: int32
          minimum: 0
        max_conns:
          type: integer
          format: int32
          minimum: 0
    BackendPoolListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                type: string
    BackendListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/BackendInfo'
    BackendInfoResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/BackendInfo'

    # -------------------------------------------------------------------------
    # Routes
    # -------------------------------------------------------------------------
...
security:
  - bearerAuth: []